refactor(canvas): split target capability workflow
- extract target capability normalization, prompts, and QApp catalog helpers

- split target workflow attachments, prompt builders, and context utilities

- add media dispatch safety coverage for target runs
